requirements.txt
annotated-types==0.7.0
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.11.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.2
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2
django-bootstrap4==22.3
django-filter==24.3
django-geojson==4.0.0
django-heroku==0.3.1
django-leaflet==0.28.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.54.1
GDAL @ file:///C:/b/abs_60gjdf6wzx/croot/gdal-split_1734448199581/work/build/swig/python
geojson==3.2.0
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.0.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
multimethod==1.12
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pandas-profiling==3.6.6
parso==0.8.4
patsy==0.5.6
phik==0.12.4
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.5
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.3
statsmodels==0.14.4
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typeguard==4.3.0
typing_extensions==4.6.3
tzdata==2022.7
urllib3==2.2.3
visions==0.7.6
wcwidth==0.2.13
whitenoise==6.4.0
wincertstore==0.2
wordcloud==1.9.3
ydata-profiling==4.11.0
zipp==3.19.2
